+++ src/main.rs
@@ -193,6 +193,8 @@ fn main() -> Result<(), Box<dyn Error>> {
         nu_engine::whole_stream_command(AnalyticsDataverses::new(state.clone())),
         nu_engine::whole_stream_command(AnalyticsIndexes::new(state.clone())),
         nu_engine::whole_stream_command(AnalyticsLinks::new(state.clone())),
+        nu_engine::whole_stream_command(AnalyticsBuckets::new(state.clone())),
+        nu_engine::whole_stream_command(AnalyticsPendingMutations::new(state.clone())),
         nu_engine::whole_stream_command(Buckets {}),
         nu_engine::whole_stream_command(BucketsConfig::new(state.clone())),
         nu_engine::whole_stream_command(BucketsCreate::new(state.clone())),