+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All user visible changes to this project will be documented in this file.
  - Added examples to: doc get
  - handle config syntax errors gracefully and log them nicely on startup
  - Added the `analytics links` command
+ - Added the `analytics buckets` command
+ - Added the `analytics pending-mutations` command
 
 ## 1.0.0-alpha.1 - 2021-05-20